Add allowParseLastLineAsExpression
This has 2 uses:
 - It can simplfy the REPL
- It allows us to add interpreter tests that can also execute statements without needing to touch the existing tests.

allowParseLastLineAsExpression sets the parser to a mode where it allows to parse a StmtExpression without the semicolon if its the last in the file.

This means just “1 + 1” which would normally need to be “1 + 1;” parses just fine with this flag set.

I also modified the interpreter so that the visitor for BlockStmt and StmtExpression doesn’t return Unit but instead the expression evaluted.